+++ src/components/Home/ClickableImage/clickableImage.styled.js
@@ -12,7 +12,6 @@ export const PersonImgWrapper = styled("div")`
     border-radius: 50%;
     translate: -50% -50%;
     z-index: ${({ $inlarge }) => $inlarge && 10};
-    /* transition: all 200ms ease-out; */
 
     & img {
         width: auto;
